Check that find_by method works with bad configuration
We do not support the configuration where the same column is present
both on the model table and on the translation table. However, this
exists out there because the migration methods leave the original model
column when migrating, so this is a problem of our own creation.

Since we have already implemented a change to fix issues related to this
configuration in #449, we should at least have a test to check for regressions.