Use TCP connection during testing via real DB
The thing is that by default MySQL and PSQL use unix socket to connect
to databases, which doesn't suit if somebody uses remote
(or containerized) DBs for testing. This commit makes it possible
to connect DBs via TCP connection instead, ignoring unix sockets.

Also, MySQL connection seems not working with "hostname" for Docker. mysql gem can't resolve "hostname" and still tries to use socket connection, which is not acceptable in case of using Docker.