Prefer forwarding shorthand syntax on Ruby 2.7 and later
In rails/rails#38034 the `save` definition in `ActiveRecord::Validations`
was changed from an options hash to keyword arguments to prevent warnings
on Ruby 2.7. To prevent future changes to method signatures causing
issues use the forwarding shorthand syntax available in 2.7 and later.

We have to class_eval the method because even though it's inside the
conditional, Ruby < 2.7 still checks the syntax and it raises an error.